Added new easier navigation buttons for both small and bigger screens. Created compontent ghostbuttons, which is used for WizardNavPair (larger screens for now) and in the new (#158)
WizardNavigationFooter for mobile and MD screens.
Footer is placed in the main wizard compontent, allowing users to move forward and backwards in both substeps and major steps in one. The pairs, for desktop, are renderd
in the same places as before.
The substep is now being delivered to the setupwizard component for usage.
Refactored how navigation works, making a function within setup wizard that all navigation props go thru to reach the navigation hook.
Refactored the savedata for state, before a failed call to the API overwrite the wizardstate, leading to dataloss. Now a failed API call notifies the user and keeps the state.
State is only overwritten when a successful call to the API is done.