feat(email): Implement core Postmark email service
Implements the new IEmailService with Postmark, using IOptions<SmtpSettings> for configuration and Polly for a resilient retry policy. This provides the foundational infrastructure for sending email.     Changed name for the legacy emailservice and its interface to LegacyEmailService, will remove later after we are done

Added new email factory, EmailContentFactory, that replaces PreparationService. This class now handles and creates different types of emails. Legacy IF and class has been renamed to ILegacyEmailPreparationService.